requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
APScheduler==3.10.4
aiosignal==1.3.1
anyio==4.4.0
asyncio==3.4.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
TgCrypto==1.2.5
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==4.2.2
certifi==2024.2.2
cffi==1.17.0rc1
charset-normalizer==2.1.1
cryptography==42.0.7
Deprecated==1.2.13
firebase-admin==6.0.1
frozenlist==1.4.1
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.64.0
grpcio-status==1.64.0
h11==0.14.0
heroku==0.1.4
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.2
idna==3.4
magic-filter==1.0.9
msgpack==1.0.4
multidict==6.0.5
proto-plus==1.22.1
protobuf==4.21.12
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
Pyrogram==2.0.93
PySocks==1.7.1
python-dateutil==1.5
python-dotenv==0.21.0
python-telegram-bot==20.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
Telethon==1.26.1
tornado==6.4
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.14.1
yarl==1.9.4
